requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-cors-headers==3.10.1
django-filter==21.1
django-oauth-toolkit==1.6.1
djangorestframework==3.13.1
drf-social-oauth2==1.1.3
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
wrapt==1.13.3
